I suspect we need to upgrade both libtommath
and libtomcrypt to make things work with LLVM.

The GCC compilers just...crap out with <limits>,
but LLVM only craps out on libtomcrypt. Going with LLVM.

+++ src/arch/LowLevelWindow/LowLevelWindow_MacOSX.mm
@@ -428,8 +428,7 @@ RString LowLevelWindow_MacOSX::TryVideoMode( const VideoModeParams& p, bool& new
 	
 	if( bChangeVsync )
 	{
-		long swap = p.vsync ? 1 : 0;
-
+		GLint swap = p.vsync ? 1 : 0;
 		[m_Context setValues:&swap forParameter:NSOpenGLCPSwapInterval];
 		m_CurrentParams.vsync = p.vsync;
 	}
@@ -515,8 +514,7 @@ void LowLevelWindow_MacOSX::SetActualParamsFromMode( CFDictionaryRef mode )
 
 	if( !m_CurrentParams.windowed )
 	{
-		long swap;
-
+		GLint swap;
 		m_CurrentParams.width = CGDisplayPixelsWide( kCGDirectMainDisplay );
 		m_CurrentParams.height = CGDisplayPixelsHigh( kCGDirectMainDisplay );
 		CGLGetParameter( CGLGetCurrentContext(), kCGLCPSwapInterval, &swap );
